DUOS-1160 - Move DacUserAPI to UserService
